+++ src/KardFinancial/Notifications/Types/EarnedRewardSettledAttributes.cs
@@ -15,12 +15,6 @@ public record EarnedRewardSettledAttributes : IJsonOnDeserialized
     public required CommissionValue CommissionEarned { get; set; }
 
     /// <summary>
-    /// The timestamp of the originating transaction in ISO format
-    /// </summary>
-    [JsonPropertyName("transactionTimestamp")]
-    public DateTime? TransactionTimestamp { get; set; }
-
-    /// <summary>
     /// The display message associated to the notification
     /// </summary>
     [JsonPropertyName("message")]
@@ -49,6 +43,24 @@ public record EarnedRewardSettledAttributes : IJsonOnDeserialized
     /// </summary>
     [JsonPropertyName("cardProductId")]
     public string? CardProductId { get; set; }
+
+    /// <summary>
+    /// The timestamp of the originating transaction in ISO format
+    /// </summary>
+    [JsonPropertyName("transactionTimestamp")]
+    public DateTime? TransactionTimestamp { get; set; }
+
+    /// <summary>
+    /// The transaction ID
+    /// </summary>
+    [JsonPropertyName("transactionId")]
+    public required string TransactionId { get; set; }
+
+    /// <summary>
+    /// The amount of the originating transaction in cents
+    /// </summary>
+    [JsonPropertyName("transactionAmountInCents")]
+    public required int TransactionAmountInCents { get; set; }
 
     [JsonIgnore]
     public ReadOnlyAdditionalProperties AdditionalProperties { get; private set; } = new();